+++ src/Service/Sns/src/Exception/BatchRequestTooLongException.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace AsyncAws\Sns\Exception;
+
+use AsyncAws\Core\Exception\Http\ClientException;
+use Symfony\Contracts\HttpClient\ResponseInterface;
+
+/**
+ * The length of all the batch messages put together is more than the limit.
+ */
+final class BatchRequestTooLongException extends ClientException
+{
+    protected function populateResult(ResponseInterface $response): void
+    {
+        $data = new \SimpleXMLElement($response->getContent(false));
+        if (0 < $data->Error->count()) {
+            $data = $data->Error;
+        }
+        if (null !== $v = (($v = $data->message) ? (string) $v : null)) {
+            $this->message = $v;
+        }
+    }
+}